fix binding source

+++ ItemStockRepoPattern/View/Forms/Frm_ItemEdit.cs
@@ -1,8 +1,9 @@
-﻿using DevExpress.XtraEditors;
+﻿using System;
+using System.Windows.Forms;
+using DevExpress.XtraEditors;
+using ItemStockRepoPattern.Logic.Extension;
 using ItemStockRepoPattern.Logic.Repository;
 using ItemStockRepoPattern.Model;
-using System;
-using System.Windows.Forms;
 
 namespace ItemStockRepoPattern.View.Forms
 {
@@ -17,9 +18,12 @@ namespace ItemStockRepoPattern.View.Forms
 			InitializeComponent();
 		}
 
-		void SetData()
+		private void SetData()
 		{
-			itemModelBindingSource.DataSource = new ItemModel();
+			itemModelBindingSource.DataSource = _repository.GetByGuid(GuidHelper.SaveGuid);
+			parentGuidTextEdit.Properties.DataSource = _repository.FillLookItem();
+			parentGuidTextEdit.Properties.DisplayMember = "itemName";
+			parentGuidTextEdit.Properties.ValueMember = "itemGuid";
 		}
 
 		private void Frm_ItemEdit_Load(object sender, EventArgs e)
@@ -28,10 +32,12 @@ namespace ItemStockRepoPattern.View.Forms
 		}
 
 
-
 		private void AddMember()
 		{
-			itemModelBindingSource = null;
+			itemModelBindingSource.DataSource = new ItemModel();
+			parentGuidTextEdit.Properties.DataSource = _repository.FillLookItem();
+			parentGuidTextEdit.Properties.DisplayMember = "itemName";
+			parentGuidTextEdit.Properties.ValueMember = "itemGuid";
 		}
 
 		private void simpleButton1_Click(object sender, EventArgs e)
@@ -40,19 +46,17 @@ namespace ItemStockRepoPattern.View.Forms
 			{
 				_item = (ItemModel)itemModelBindingSource.DataSource;
 				var ch = _repository.Save(_item);
-				if (ch == 0)
+				switch (ch)
 				{
-					MessageBox.Show(@"Error");
-				}
-				else if (ch == 1)
-				{
-					MessageBox.Show(@"Done Insert");
-
-				}
-				else if (ch == 2)
-				{
-					MessageBox.Show(@"Done UPdate");
-
+					case 0:
+						MessageBox.Show(@"Error");
+						break;
+					case 1:
+						MessageBox.Show(@"Done Insert");
+						break;
+					case 2:
+						MessageBox.Show(@"Done UPdate");
+						break;
 				}
 			}
 			catch (Exception exception)
@@ -64,31 +68,38 @@ namespace ItemStockRepoPattern.View.Forms
 
 		private void simpleButton2_Click(object sender, EventArgs e)
 		{
-			AddMember();
+			try
+			{
+				AddMember();
+			}
+			catch (Exception exception)
+			{
+				Console.WriteLine(exception);
+				throw;
+			}
 		}
 
 		private void simpleButton3_Click(object sender, EventArgs e)
 		{
 			try
 			{
+				if (MessageBox.Show("are you ready to delete", "delete", MessageBoxButtons.OKCancel) !=
+				    DialogResult.OK) return;
 				_item = (ItemModel)itemModelBindingSource.DataSource;
 
-				int ch = _repository.Delete(_item.itemGuid);
-				if (ch == 0)
+				var ch = _repository.Delete(_item.itemGuid);
+				switch (ch)
 				{
-					MessageBox.Show(@"Error");
+					case 0:
+						MessageBox.Show(@"Error");
+						break;
+					case 1:
+						MessageBox.Show(@"Done Insert");
+						break;
+					case 2:
+						MessageBox.Show(@"Done UPdate");
+						break;
 				}
-				else if (ch == 1)
-				{
-					MessageBox.Show(@"Done Insert");
-
-				}
-				else if (ch == 2)
-				{
-					MessageBox.Show(@"Done UPdate");
-
-				}
-
 			}
 			catch (Exception exception)
 			{
